Dblib stopped selecting the database, with no errors or anything..
So, I had to add another, bad driver. mssql this time. This is actually even worse since it's not a PDO thingie, nor has prepare/bind. But it did give me an error message (exception) when I did wrong and it does give me the correct data when I did right.